+++ examples/minimal/src/main.rs
@@ -10,7 +10,8 @@ use ratzilla::{
     event::KeyCode, event::MouseButton, event::MouseEventKind, WebRenderer,
 };
 
-use examples_shared::{backend_from_query_param, BackendType};
+use examples_shared::{BackendType};
+use examples_shared::backend::multi_backend_builder;
 
 fn main() -> io::Result<()> {
     let counter = Rc::new(RefCell::new(0));
@@ -18,7 +19,7 @@ fn main() -> io::Result<()> {
     let mouse_button = Rc::new(RefCell::new(None::<MouseButton>));
     let mouse_event_kind = Rc::new(RefCell::new(None::<MouseEventKind>));
 
-    let (_backend_type, terminal) = backend_from_query_param(BackendType::Dom)
+    let (_backend_type, terminal) = multi_backend_builder(BackendType::Dom)
         .build_terminal()?;
 
     terminal.on_key_event({
